`/settings`を v2 に移行
Fixes #1087
