+++ app/(v2)/settings/RenameForm.tsx
@@ -1,17 +1,75 @@
 "use client";
 
+import { ResultOf } from "@graphql-typed-document-node/core";
 import { ErrorMessage } from "@hookform/error-message";
 import { zodResolver } from "@hookform/resolvers/zod";
 import clsx from "clsx";
-import { useEffect } from "react";
+import { useCallback } from "react";
 import { useForm } from "react-hook-form";
+import { useMutation } from "urql";
 import z from "zod";
 
 import Button from "~/components/Button";
 import { TextInput } from "~/components/TextInput";
 import useToaster from "~/components/Toaster/useToaster";
+import { graphql } from "~/gql";
 
-import useRename from "./useRename";
+export const RenameMutation = graphql(`
+  mutation SettingsPage_UseRename($renameTo: String!) {
+    changeUserDisplayName(renameTo: $renameTo) {
+      __typename
+      ... on ChangeUserDisplayNameInvalidNameError {
+        name
+      }
+      ... on ChangeUserDisplayNameSucceededPayload {
+        user {
+          id
+          displayName
+        }
+      }
+    }
+  }
+`);
+export const useRename = ({
+  onSucceeded,
+}: {
+  onSucceeded(
+    data: Extract<
+      ResultOf<typeof RenameMutation>["changeUserDisplayName"],
+      { __typename: "ChangeUserDisplayNameSucceededPayload" }
+    >
+  ): void;
+}) => {
+  const [{ fetching }, mutate] = useMutation(RenameMutation);
+
+  const rename = useCallback(
+    async (renameTo: string) => {
+      const result = await mutate({ renameTo });
+      if (result.error || !result.data) {
+        return;
+      }
+      switch (result.data.changeUserDisplayName.__typename) {
+        case "ChangeUserDisplayNameInvalidNameError": {
+          break;
+        }
+        case "ChangeUserDisplayNameSucceededPayload": {
+          onSucceeded(result.data.changeUserDisplayName);
+          break;
+        }
+      }
+    },
+    [onSucceeded, mutate]
+  );
+  return [{ fetching }, rename] as const;
+};
+
+const formSchema = z.object({
+  renameTo: z
+    .string()
+    .min(1, { message: "1文字以上の名前にしてください。" })
+    .max(20, "20文字以下の名前にしてください。"),
+});
+type FormSchema = z.infer<typeof formSchema>;
 
 export default function RenameForm({
   className,
@@ -20,29 +78,18 @@ export default function RenameForm({
   className?: string;
   style?: React.CSSProperties;
 }) {
-  const formSchema = z.object({
-    renameTo: z
-      .string()
-      .min(1, { message: "1文字以上の名前にしてください。" })
-      .max(20, "20文字以下の名前にしてください。"),
-  });
-  type FormSchema = z.infer<typeof formSchema>;
-
   const {
     register,
     handleSubmit,
     reset,
-    formState: { errors, isSubmitSuccessful },
+    formState: { errors },
   } = useForm<FormSchema>({
     resolver: zodResolver(formSchema),
   });
-  useEffect(() => {
-    if (isSubmitSuccessful) reset();
-  }, [reset, isSubmitSuccessful]);
-
   const callToast = useToaster();
   const [{ fetching }, rename] = useRename({
     onSucceeded({ user: { displayName } }) {
+      reset();
       callToast(
         <>
           名前を<strong className={clsx("font-bold")}>{displayName}</strong>
@@ -56,32 +103,33 @@ export default function RenameForm({
     <form
       onSubmit={handleSubmit(({ renameTo }) => rename(renameTo))}
       style={style}
-      className={clsx(className, "flex items-start gap-x-2")}
+      className={clsx(className, "flex items-start justify-between gap-x-2")}
     >
-      <div className={clsx(["relative"])}>
+      <div className={clsx("relative flex flex-col gap-y-1")}>
         <TextInput
           {...register("renameTo")}
           disabled={fetching}
-          className={clsx(["w-72 px-2 py-1"])}
           placeholder="新しい表示名"
+          className={clsx("w-96")}
         />
         <ErrorMessage
           errors={errors}
           name="renameTo"
           render={({ message }) => (
-            <p className={clsx("mt-1 text-xs text-red-400")}>{message}</p>
+            <p className={clsx("text-xs text-error-primary")}>{message}</p>
           )}
         />
       </div>
-      <Button
-        submit
-        className={clsx("shrink-0")}
-        color="blue"
-        size="small"
-        icon="plus"
-        text="変更"
-        disabled={!fetching}
-      />
+      <div className={clsx("flex shrink-0")}>
+        <Button
+          submit
+          color="blue"
+          size="small"
+          icon="plus"
+          text="変更"
+          disabled={fetching}
+        />
+      </div>
     </form>
   );
 }
