--- app/(v2)/settings/layout.tsx
@@ -1,9 +1,0 @@
-import clsx from "clsx";
-
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return (
-    <main className={clsx("container mx-auto max-w-screen-lg py-8")}>
-      {children}
-    </main>
-  );
-}
